move db ops to utils
